refactor(weather): Update forecast model imports in view model layer
- Adjusted imports in city_search_state, city_weather_notifier, weather_view_model_state,
and related providers to match the new model location under lib/features/weather/models/.
- Maintains consistency with the updated feature-based directory structure.
