+++ lib/view_model/city_search_state.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../models/forecast.dart';
+import '../features/weather/models/forecast.dart';
 
 part 'city_search_state.freezed.dart';
